### 2025-01-05 8a10248
added doi's to paper
