+++ paper/paper.bib
@@ -7,11 +7,12 @@
   publisher={North Holland}
 }
 
-@book{kaufman_rousseeuw_2009,
+@book{kaufman_rousseeuw_2005,
   title={Finding groups in data: an introduction to cluster analysis},
   author={Kaufman, Leonard and Rousseeuw, Peter J},
-  year={2009},
-  publisher={John Wiley \& Sons}
+  year={2005},
+  publisher={John Wiley \& Sons},
+  doi={10.1002/9780470316801}
 }
 
 @inproceedings{krishnapuram_1999,
@@ -21,7 +22,8 @@
   volume={3},
   pages={1281--1286},
   year={1999},
-  organization={IEEE}
+  organization={IEEE},
+  doi={10.1109/fuzzy.1999.790086}
 }
 
 @book{mcneil_1977,
@@ -39,7 +41,8 @@
   number={5},
   pages={1003--1016},
   year={2002},
-  publisher={IEEE}
+  publisher={IEEE},
+  doi={10.1109/tkde.2002.1033770}
 }
 
 @Manual{R_cluster,
@@ -47,7 +50,7 @@
   author = {Martin Maechler and Peter Rousseeuw and Anja Struyf and Mia Hubert and Kurt Hornik},
   year = {2022},
   url = {https://CRAN.R-project.org/package=cluster},
-  note = {R package version 2.1.4 --- For new features, see the 'Changelog' file (in the package source)},
+  note = {R package version 2.1.4 --- For new features, see the 'Changelog' file (in the package source)}
 }
 
 @Manual{R_fastkmedoids,
@@ -55,7 +58,7 @@
   author = {Xun Li},
   year = {2021},
   note = {R package version 1.2},
-  url = {https://CRAN.R-project.org/package=fastkmedoids},
+  doi = {10.32614/cran.package.fastkmedoids}
 }
 
 @Article{R_fclust,
@@ -64,7 +67,8 @@
   journal = {The {R} Journal},
   year = {2019},
   volume = {11},
-  url = {https://journal.r-project.org/archive/2019/RJ-2019-017/RJ-2019-017.pdf},
+  number = {1},
+  doi = {10.32614/RJ-2019-017}
 }
 
 @Manual{R_qtcat,
@@ -72,7 +76,7 @@
   author = {Jonas Klasen},
   year = {2022},
   note = {R package version 0.4.3},
-  url = {http://github.com/QTCAT/qtcat},
+  url = {http://github.com/QTCAT/qtcat}
 }
 
 @Manual{R_vegclust,
@@ -82,5 +86,5 @@
   year = {2010},
   volume = {21},
   pages = {1138-1151},
-  doi = {10.1111/j.1654-1103.2010.01211.x},
+  doi = {10.1111/j.1654-1103.2010.01211.x}
 }
