### 2025-01-06 ea81917
substantially shortened the paper draft
