+++ paper/paper.md
@@ -27,16 +27,16 @@ bibliography: paper.bib
 
 # Summary
 
-Cluster analysis identifies optimal groupings of observations that share similar
+Cluster analysis identifies groupings of observations that share similar
 characteristics.
-One popular approach is to use medoid-based methods where each cluster center is
+One popular approach are medoid-based methods where each cluster center is
 represented by one *typical* observation [@kaufman_rousseeuw_2005].
 The R package `fuzzyclara` makes a wide range of clustering algorithms conveniently available.
-It not only covers classical *hard clustering* methods (where one observation belongs to exactly one cluster)
-and alternative *fuzzy clustering* methods (where the characteristics of each observation
-are shaped by its partial membership to different clusters), but specifically also implements
-the option to combine such algorithms with subsampling-based estimation techniques to make the
-estimation on large data feasible.
+It covers classical *hard clustering* methods (where one observation belongs to exactly one cluster)
+and alternative *fuzzy clustering* methods (where each observation
+is shaped by its partial membership to different clusters), and further implements
+the option to combine such algorithms with subsampling-based estimation techniques
+to make the estimation on large data feasible.
 The package additionally provides convenience functionalities and visualization
 techniques to cover the whole workflow for real-world clustering applications.
 
@@ -48,27 +48,22 @@ of a set of observations based on a predefined number of clusters.
 Medoid-based versions of this strategy build clusters based on *medoids*,
 i.e. one observation per cluster best representing its typical characteristics.
 The most prominent representative of medoid-based clustering is the
-*partitioning around medoids* (PAM) algorithm, which is robust and applicable to
-many data situations [@kaufman_rousseeuw_2005].
+*partitioning around medoids* (PAM) algorithm [@kaufman_rousseeuw_2005].
 
 The PAM algorithm, however, has two drawbacks.
-First, its estimation is often only hardly or not at all feasible in large data
-situations with thousands of observations.
-The algorithm requires the computation of a (dis)similarity matrix between all
-observations, scaling quadratically ($O(n^2)$) in terms of runtime and
-memory usage.
-Sampling-based algorithms such as CLARA [@kaufman_rousseew_1986] or
-CLARANS [@ng_han_2002] make the estimation feasible in such situations.
-Second, PAM is a hard clustering algorithm where each observation is rigidly
-assigned to a single cluster.
-This assumption is not best resembling reality in many data situations where
-observations may share characteristics of several *typical* clusters.
-Such structures are taken into account by *fuzzy clustering* methods which
+First, its estimation can become hardly feasible in data situations with thousands of observations,
+scaling quadratically ($O(n^2)$) in terms of runtime and memory usage.
+Sampling-based algorithms like CLARA [@kaufman_rousseew_1986] or
+CLARANS [@ng_han_2002] allow for a more efficient estimation in such situations.
+Second, PAM is a hard clustering algorithm where each observation is assigned to a single cluster.
+This assumption is often not resembling reality when observations share
+characteristics of several *typical* clusters.
+Such structures are taken into account by *fuzzy clustering* methods, which
 compute membership scores for each observation to every cluster.
 
 The statistical software R already provides a wide range of packages implementing
 clustering algorithms for large or fuzzy data.
-The package `cluster` [@R_cluster] contains diverse clustering routines
+The `cluster` package [@R_cluster] contains diverse clustering routines
 developed by @kaufman_rousseeuw_2005 including the CLARA algorithm for large
 data and the FANNY algorithm for fuzzy data.
[…]
