+++ pkgdown.yml
@@ -4,4 +4,4 @@ pkgdown_sha: ~
 articles:
   main_functionality: main_functionality.html
   vignette_shiny: vignette_shiny.html
-last_built: 2025-01-05T10:33Z
+last_built: 2025-01-06T15:16Z
