### 2025-01-06 #7609 @bauer-alex
Thanks upfront to both of you for your efforts. Highly appreciated!

I now shortened the manuscript by over 250 words and integrated some reference doi's that editorialbot was hinting to. I hope you guys can work with the current manuscript, even if it's 100-150 words over the 1000 word limit. If you would be ok with this approach, I would first await the revision comments before further shortening it. But I can of course already now shorten it to 1000-1050 words if you prefer that.
