### 2025-03-10 #7609 @kanishkan91
@amoghantarkar based on our email exchanges, would you be interested in reviewing this paper for JOSS?
