### 2025-03-11 #7887 @amoghantarkar
@kanishkan91 @bauer-alex Started peer review. The license doesn't seem like a valid MIT or BSD2 license and implies not open source as it mentions: COPYRIGHT HOLDER: Maximilian Weigert. Can you please check or comment
