+++ pkgdown.yml
@@ -4,4 +4,4 @@ pkgdown_sha: ~
 articles:
   main_functionality: main_functionality.html
   vignette_shiny: vignette_shiny.html
-last_built: 2025-03-12T09:03Z
+last_built: 2025-03-12T09:44Z
