### 2025-03-12 #7887 @bauer-alex
Thank you for the note. I now corrected the license to an MIT license, see the new `LICENSE.md` file. Had to keep the other `LICENSE` file, otherwise `devtools::check()` would throw a warning. Since also other packages have this setup (e.g. [the coalitions R package](https://github.com/adibender/coalitions)) I hope this solution is fine for you.
