### 2025-03-13 #7887 @amoghantarkar
@kanishkan91 @bauer-alex
for checklist: Installation instructions: Is there a clearly-stated list of dependencies?
The dependencies could be listed. I ran into errors had to add install.packages(c("vegclust", "ggwordcloud", "ggpubr", "factoextra")) but if all required packages could be listed would be better in the read.md
