### 2025-03-13 #7887 @amoghantarkar
@bauer-alex Functionality documentation overall looks good but wanted to make citation addition suggestion.  I noticed  marked as TODO in the fuzzyclara.R code around CLARA and CLARANS algorithms. Consider adding a CITATION.cff file  similar to https://github.com/VeinsOfTheEarth/rabpro/blob/main/CITATION.cff

cc @kanishkan91
