### 2025-03-13 #7887 @amoghantarkar
@bauer-alex Community guidelines: meets basic needs but could consider enhancing the contribution section with more details about R code style and adopt one. cc @kanishkan91
