+++ vignettes/vignette_shiny.Rmd
@@ -29,7 +29,6 @@ library(shiny)
 
 ### Hard clustering
 ```{r, hard clustering example}
-# TODO replace USArrests by a better dataset (maybe TourIST)
 cc_hard <- fuzzyclara(data        = USArrests,
                        clusters    = 3,
                        metric      = "euclidean",
