### 2025-03-19 1a2cfaa
added note on R style guide to readme
