+++ tests/testthat/test-assign_cluster.R
@@ -5,9 +5,10 @@ test_that("assign_cluster", {
   row.names(data) <- data$Name
 
   # hard
-  result <- assign_cluster(data = data,
-                           medoids = c("Alabama", "Alaska", "Arizona"),
-                           metric = "Euclidean", type = "hard",
+  result <- assign_cluster(data              = data,
+                           medoids           = c("Alabama", "Alaska", "Arizona"),
+                           metric            = "Euclidean",
+                           type              = "hard",
                            return_distMatrix = TRUE)
 
   # check whole object
@@ -16,9 +17,11 @@ test_that("assign_cluster", {
   expect_identical(dim(result$distance_to_medoids), as.integer(c(nrow(data), 3)))
 
   # fuzzy
-  result <- assign_cluster(data = data,
-                           medoids = c("Alabama", "Alaska", "Arizona"),
-                           metric = "Euclidean", type = "fuzzy", m = 3,
+  result <- assign_cluster(data              = data,
+                           medoids           = c("Alabama", "Alaska", "Arizona"),
+                           metric            = "Euclidean",
+                           type              = "fuzzy",
+                           m                 = 3,
                            return_distMatrix = TRUE)
 
   # check whole object
@@ -28,9 +31,12 @@ test_that("assign_cluster", {
 
   # with distance matrix
   dist <- as.matrix(proxy::dist(data[, -1]))
-  result <- assign_cluster(data = data, dist_matrix = dist,
-                           medoids = c("Alabama", "Alaska", "Arizona"),
-                           metric = "Euclidean", type = "fuzzy", m = 3,
+  result <- assign_cluster(data              = data,
+                           dist_matrix       = dist,
+                           medoids           = c("Alabama", "Alaska", "Arizona"),
+                           metric            = "Euclidean",
+                           type              = "fuzzy",
+                           m                 = 3,
                            return_distMatrix = TRUE)
 
   # check whole object
@@ -44,8 +50,8 @@ test_that("assign_cluster", {
 test_that("calculate_memb_score", {
 
   data(USArrests)
-  data <- USArrests %>% tibble::rownames_to_column(var = "Name")
-  medoids <- c("Alabama", "Alaska", "Arizona")
+  data         <- USArrests %>% tibble::rownames_to_column(var = "Name")
+  medoids      <- c("Alabama", "Alaska", "Arizona")
   data_medoids <- data %>% filter(Name %in% medoids)
 
   dist_dat <- proxy::dist(x         = data[, -1],
@@ -63,6 +69,5 @@ test_that("calculate_memb_score", {
   expect_identical(unname(unlist(memb2)), c(1, 0, 0))
   memb3 <- calculate_memb_score(dist_med = c(1, 1, 1), m = 2) # same distance to all medoids
   expect_identical(round(unname(unlist(memb3))), round(rep(1/3, 3)))
-
 
 })
