+++ tests/testthat/test-plot_fuzzyclara.R
@@ -189,39 +189,61 @@ test_that("plot_fuzzyclara_clara", { # plots
   
   
   # silhouette, fuzzy, silhouette_subsample = TRUE
-  invisible(capture.output(p <- plot(x = cc_fuzzy, data = USArrests, type = "silhouette",
-                                     silhouette_subsample = TRUE, silhouette_table = FALSE)))
+  invisible(capture.output(p <- plot(x                    = cc_fuzzy,
+                                     data                 = USArrests,
+                                     type                 = "silhouette",
+                                     silhouette_subsample = TRUE,
+                                     silhouette_table     = FALSE)))
   # check class of object
   expect_s3_class(p, "ggplot")
   expect_s3_class(p$layers[[1]]$geom, "GeomBar")
   
   
   # focus scatterplot
-  p <- plot(x = cc_fuzzy, data = USArrests_enriched, type = "scatterplot",
-            x_var = "Murder", y_var = "Assault", focus = TRUE, focus_clusters = c(1, 2))
+  p <- plot(x              = cc_fuzzy,
+            data           = USArrests_enriched,
+            type           = "scatterplot",
+            x_var          = "Murder",
+            y_var          = "Assault",
+            focus          = TRUE,
+            focus_clusters = c(1, 2))
   # check class of object
   expect_s3_class(p, "ggplot")
   expect_s3_class(p$layers[[1]]$geom, "GeomPoint")
   
-  expect_error(plot(x = cc_fuzzy, data = USArrests_enriched, type = "scatterplot",
-                    x_var = "Murder", y_var = "Assault",
-                    focus = TRUE, focus_clusters = c(0, 2)))
+  expect_error(plot(x              = cc_fuzzy,
+                    data           = USArrests_enriched,
+                    type           = "scatterplot",
+                    x_var          = "Murder",
+                    y_var          = "Assault",
+                    focus          = TRUE,
+                    focus_clusters = c(0, 2)))
   
   # focus pca
-  p <- plot(x = cc_fuzzy, data = USArrests_enriched, type = "pca",
-            focus = TRUE, focus_clusters = c(1, 2))
+  p <- plot(x              = cc_fuzzy,
+            data           = USArrests_enriched,
+            type           = "pca",
+            focus          = TRUE,
+            focus_clusters = c(1, 2))
   # check class of object
   expect_s3_class(p, "ggplot")
   expect_s3_class(p$layers[[1]]$geom, "GeomPoint")
   
-  p <- plot(x = cc_fuzzy, data = USArrests_enriched, type = "pca",
-            focus = TRUE, focus_clusters = c(1, 2), group_by = "Area")
+  p <- plot(x              = cc_fuzzy,
+            data           = USArrests_enriched,
+            type           = "pca",
+            focus          = TRUE,
+            focus_clusters = c(1, 2),
+            group_by       = "Area")
   # check class of object
   expect_s3_class(p, "ggplot")
   expect_s3_class(p$layers[[1]]$geom, "GeomPoint")
   
-  expect_error(plot(x = cc_fuzzy, data = USArrests_enriched, type = "pca",
-                    focus = TRUE, focus_clusters = c(0, 2)))
+  expect_error(plot(x              = cc_fuzzy, 
+                    data           = USArrests_enriched,
+                    type           = "pca",
+                    focus          = TRUE, 
+                    focus_clusters = c(0, 2)))
   
   
 })
@@ -259,80 +281,113 @@ test_that("plot_fuzzyclara_clarans", { # plots
                                          "Northeast")))
   
   ## error for wrong variable specification
-  expect_error(plot(x = cc_hard, data = USArrests_enriched, variable = "Assalt"))
[…]
