### 2025-03-27 db55259
function documentations: added CLARA and CLARANS references
