+++ man/clustering_local.Rd
@@ -60,5 +60,7 @@ a hard or fuzzy way. The function can either be called using a common
 dissimilarity metric or a self-defined distance function.
 }
 \references{
-TODO add CLARANS paper
+Ng, R. T., and Han, J. (2002). CLARANS: A method for clustering
+objects for spatial data mining. \emph{IEEE transactions on knowledge and data engineering},
+14(5), 1003--1016. \doi{10.1109/tkde.2002.1033770}.
 }
