+++ man/clustering_clarans.Rd
@@ -67,5 +67,7 @@ If the clustering is run on mulitple cores, the verbose messages are printed
 in a file \code{clustering_progress.log} (if \code{verbose > 0}).
 }
 \references{
-TODO add CLARANS paper
+Ng, R. T., and Han, J. (2002). CLARANS: A method for clustering
+objects for spatial data mining. \emph{IEEE transactions on knowledge and data engineering},
+14(5), 1003--1016. \doi{10.1109/tkde.2002.1033770}.
 }
