### 2025-03-27 #7887 @bauer-alex
@amoghantarkar Thank you for all these valuable comments. I now went through everything. Here one compact response to all your comments:

1. Dependencies
> for checklist: Installation instructions: Is there a clearly-stated list of dependencies?
Dependencies are correctly listed in the DESCRIPTION file but Readme.md it the specialized and key dependencies could be listed. I ran into errors and had to add specialized packages like install.packages(c("vegclust", "ggwordcloud", "ggpubr", "factoextra")) and could be added in the README.md before the main installation step on fuzzy cluster installation. let me know this once!

**Response**
I tried to find out why the automatic installation of the package dependencies didn't work, but I unfortunately couldn't reproduce it. For me, both on Windows 11 and on Ubuntu 24.10 (specifically Linux Mint 22) the automatic installation did work. Rather than adding documentation on this in the README -- *if possible* -- I would like to resolve this issue since the automatic installation should actually work (source: [r-pkgs.org](https://r-pkgs.org/dependencies-in-practice.html#confusion-about-imports)) Could you let me know which operating system you installed the package on, and maybe the R version and the prompt you used for installation (`devtools::install_github()` I assume)? If we cannot resolve it I'm happy to go with your suggestion and add some extra documentation to the README of course.

2. Documentation
> Functionality documentation overall looks good but wanted to make citation addition suggestion. I noticed marked as TODO in the fuzzyclara.R code around CLARA and CLARANS algorithms. Consider adding a CITATION.cff file similar to https://github.com/VeinsOfTheEarth/rabpro/blob/main/CITATION.cff

**Response**
Sorry for having overlooked the `TODO` comments in our code. Stupid mistake. We now resolved all `TODO` comments and added the CLARA and CLARANS references in the function documentations.
Regarding the `CITATION.cff` file: Good idea. We would add such a citation file once the paper is properly published if that's ok with you?

3. R code style
> Community guidelines: meets basic needs but could consider enhancing the contribution section with more details about R code style and adopt one.

**Response**
Our coding style was indeed somewhat inconsistent. We now added a link to the [Advanced R style guide](http://adv-r.had.co.nz/Style.html) in the `How to contribute` section in the README. Additionally, I very thoroughly worked through all the code again to make sure that the style guide is consistently applied, substantially changing our comment style, the indentations, etc. [See the respective commit](https://github.com/bauer-alex/fuzzyclara/commit/a28a81fc2379dbef575d1284e2626adff37f84b2) as reference that this code overhaul happened.

4. Final comment
> I ran the examples and tested functionality and overall looks good to me. On installation dependencies i have a review comment and few suggestions in other checklist items please check and let me know and i can approve on my end. Thanks!

**Response**
We assume this was rather a summary comment about all the previously given notes? If not please let us know.
