### 2025-03-30 #7887 @amoghantarkar
Hi @bauer-alex Please see OS and R version details! I did clean R Studio setup on Mac OS 15.
If you have existing libraries installed you might not get the error and it might working in Windows Linux. If you unable to test on Mac OS or setup a clean R studio to reproduce the issue let us know you can consider adding in documentation. I will check it off since it is your working on Windows, Linux with the comment. fyi @kanishkan91

OS:

ProductName:		macOS
ProductVersion:		15.0 Sequoia
BuildVersion:		24A335

> version
               _
platform       x86_64-apple-darwin20
arch           x86_64
os             darwin20
system         x86_64, darwin20
status
major          4
minor          4.3
year           2025
month          02
day            28
svn rev        87843
language       R
version.string R version 4.4.3 (2025-02-28)
nickname       Trophy Case
