### 2025-04-01 #7887 @kanishkan91
@bauer-alex and @amoghantarkar thanks for keeping this review going.

@bauer-alex - One thing to consider is to add a MAC build workflow in your GitHub workflows (This would only require editing the yaml files I think). Its a relatively easy change to make and would ensure that the package builds correctly across several environments.
