### 2025-04-01 #7887 @kanishkan91
@hahsan1 would you be interested in reviewing this paper for JOSS?
