### 2025-05-03 #7887 @hahsan1
I have completed the review checklist. Overall, this is a well-developed package with appropriate documentation and functionality. Some minor comments below.

The manuscript is clear and well-structured. Just a couple of suggested fixes:
Original: “This assumption is often not resembling reality...”
Suggested: “This assumption often does not reflect reality...”

Original: “depicting the minial average weighted distance”
Suggested: “depicting the minimal average weighted distance”

The package vignette is a useful demonstration of `fuzzyclara`'s functionality. However, it’s a little bare-bones. I think the examples could be improved by clarifying the outputs and adding some interpretations. For example, explain the role of medoids and what the minimum average distance signifies. Basically, what insights can be drawn from the clustering in this example. This would be helpful for new users. (This is more of a future "to do")
