### 2025-05-05 #7887 @kanishkan91
@hahsan1 Thanks so much!

@bauer-alex as you respond to the comments above, could you also respond regarding the status of the issue below that the first reviewer raised?

> [@bauer-alex](https://github.com/bauer-alex) - One thing to consider is to add a MAC build workflow in your GitHub workflows (This would only require editing the yaml files I think). This would be in addition to the CMD check you already have. Its a relatively easy change to make and would ensure that the package builds correctly across several environments.
