### 2025-05-16 4eabf60
minor typo fixes in paper
