### 2025-05-20 a044da7
vignette update + some very minor comment corrections
