### 2025-05-20 #7887 @bauer-alex
Sorry for the delay. We will work through the last remaining open MAC build comment in the following days and give an update. For now, I resolved all other remaining comments of @hahsan1, specifically:

> Original: “This assumption is often not resembling reality...” Suggested: “This assumption often does not reflect reality...”

-> Resolved

> Original: “depicting the minial average weighted distance” Suggested: “depicting the minimal average weighted distance”

-> Resolved

> The package vignette is a useful demonstration of `fuzzyclara`'s functionality. However, it’s a little bare-bones. I think the examples could be improved by clarifying the outputs and adding some interpretations. For example, explain the role of medoids and what the minimum average distance signifies. Basically, what insights can be drawn from the clustering in this example. This would be helpful for new users. (This is more of a future "to do")

-> This was a very valid comment. I now added explanatory texts to all parts of the vignette, s.t. it is better readable.
