### 2025-05-21 #7887 @kanishkan91
👋 @bauer-alex  - Thanks! I'm taking a read through the latest manuscript. After you resolve the MAC comment, we need to set up the archive for your new release.

So as next steps for you:

- [ ] Obvious remaining step is addressing the MAC build

- [ ] Check the archival deposit (e.g., in Zenodo) to ensure it has the correct metadata. This includes the title (should match the paper title) and author list (make sure the list matches the people and order in your paper). You may also add the authors' ORCID.

- [ ] Check that software archive exists and title and author list match JOSS paper (or purposefully do not).

- [ ] Please respond with the 1) version and 2) DOI of the archived version here

Huge thank you to @amoghantarkar  and @hahsan1 for a very detailed and constructive review. JOSS is volunteer run and would not be able to function without reviewers like yourself!
