### 2025-05-26 959f5e9
added MacOS installation note to readme
