+++ pkgdown.yml
@@ -1,7 +1,7 @@
 pandoc: 3.1.11
-pkgdown: 2.1.2
+pkgdown: 2.1.3
 pkgdown_sha: ~
 articles:
   main_functionality: main_functionality.html
   vignette_shiny: vignette_shiny.html
-last_built: 2025-05-20T06:50Z
+last_built: 2025-05-26T19:37Z
