### 2025-05-26 #7887 @bauer-alex
**Regarding the remaining open MacOS issues**

First, regarding the installation problems experienced by @amoghantarkar on MacOS, we now tried to reproduce the issues on two R setups:
1. an M4 Mac, with MacOS 15.4.1, running R 4.4.3
2. an M1 Mac, with MacOS 11.7.10, running R 4.5.0 (completely clean R setup)

On both systems the installation worked without any issue.
Accordingly, I now simply added a note (with [this commit](https://github.com/bauer-alex/fuzzyclara/commit/959f5e9ec21b520b80d3d80ad2b5596efed75b60)) to the package readme to make people aware of the potential issue. The note was based on [this above comment](https://github.com/openjournals/joss-reviews/issues/7887#issuecomment-2720349439) here in the revision chat.

Second, regarding the additional build workflow for MacOS (which was based on [this above comment](https://github.com/openjournals/joss-reviews/issues/7887#issuecomment-2767925772)), I checked this now and actually the MacOS build is already included in our R CMD CHECK workflow, so there's nothing to resolve as far as I see it. See [this line in our R-CMD-check.yaml file](https://github.com/bauer-alex/fuzzyclara/blob/959f5e9ec21b520b80d3d80ad2b5596efed75b60/.github/workflows/R-CMD-check.yaml#L21).

I hope @kanishkan91 you agree with this approach?
