### 2025-05-27 #7887 @kanishkan91
@bauer-alex while the track editor takes a look at this paper- could you fix this missing DOI-

10.2307/2346743 may be a valid DOI for title: Interactive data analysis: a practical primer
