### 2025-05-27 2aee5b1
paper: removed deprecated reference in bib
