+++ paper/paper.bib
@@ -26,13 +26,6 @@
   doi={10.1109/fuzzy.1999.790086}
 }
 
-@book{mcneil_1977,
-  title={Interactive data analysis: a practical primer},
-  author={McNeil, Donald Raymond},
-  year={1977},
-  publisher={New York: Wiley}
-}
-
 @article{ng_han_2002,
   title={CLARANS: A method for clustering objects for spatial data mining},
   author={Ng, Raymond T. and Han, Jiawei},
