### 2025-06-03 #7887 @kanishkan91
@bauer-alex I see the paper here- https://joss.theoj.org/papers/10.21105/joss.07887
